apps/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "FastAPI gateway with JWT auth and RBAC"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pyjwt[crypto]>=2.8.0",
    "bcrypt>=4.1.0",
    "cryptography>=42.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "structlog>=24.0.0",
    "grpcio>=1.62.0",
    "protobuf>=4.25.0",
    "pwdlib[argon2]>=0.3.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/api_gateway"]
[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
